+++ backend/src/services/chat.ts
